# Optimize SipHash when msg[D.low..D.high] lives on one locale