--- a/src/SipHash.chpl
+++ b/src/SipHash.chpl
@@ -1,4 +1,5 @@
 module SipHash {
+  private use AryUtil;
   
   param cROUNDS = 2;
   param dROUNDS = 4;
@@ -36,6 +37,18 @@ module SipHash {
             (p[D.low+7]: uint(64) << 56));
   }
 
+  private inline proc U8TO64_LE(p: c_ptr(uint(8)), D): uint(64) {
+    return ((p[0]: uint(64)) |
+            (p[1]: uint(64) << 8) |
+            (p[2]: uint(64) << 16) |
+            (p[3]: uint(64) << 24) |
+            (p[4]: uint(64) << 32) |
+            (p[5]: uint(64) << 40) |
+            (p[6]: uint(64) << 48) |
+            (p[7]: uint(64) << 56));
+  }
+
+
   private inline proc byte_reverse(b: uint(64)): uint(64) {
     var c: uint(64);
     c |= (b & 0xff) << 56;
@@ -55,10 +68,15 @@ module SipHash {
   }
 
   proc sipHash128(msg: [] uint(8), D): 2*uint(64) {
+    if contiguousIndices(msg) {
+      if msg[D.low].locale.id == here.id && msg[D.high].locale.id == here.id {
+        return computeSipHash(c_ptrTo(msg[D.low]), 0..#D.size, 16);
+      }
+    }
     return computeSipHash(msg, D, 16);
   }
   
-  private proc computeSipHash(msg: [] uint(8), D, param outlen: int) {
+  private proc computeSipHash(msg, D, param outlen: int) {
     if !((outlen == 8) || (outlen == 16)) {
       compilerError("outlen must be 8 or 16");
     }