Use c_ptr to optimize indexing when msg[D.low..D.high] is local. This optimizes out more complex indexing operations, particularly in `U8TO64_LE`.

If msg[D.low..D.high] is remote, but all lives on one locale, use bulk comm to localize and then use c_ptr. The old version would do a 1-byte GET for each element we needed and this uses a single GET. This adds an allocation (which we might be able to optimize out with scratch space), but the benefit of being able to use bulk comm (and then using c_ptr) far outweighs the allocation cost.

Here is 32 node performance on Aries and InfiniBand (same machines as #231)

    ./SipHashSpeedTest -nl32 --INPUTSIZE=64 --NINPUTS=10_000_000_000 --LENGTHS=fixed

| config | before   | after    |
| ------ | -------- | -------- |
| Aries  | 190 GB/s | 515 GB/s |
| FDR IB | 195 GB/s | 545 GB/s |


    ./SipHashSpeedTest -nl32 --INPUTSIZE=64 --NINPUTS=10_000_000_000 --LENGTHS=variable
(I used a 100x smaller `--NINPUTS` for "before" to finish in a reasonable amount of time.)

| config | before   | after    |
| ------ | -------- | -------- |
| Aries  |   2 GB/s | 225 GB/s |
| FDR IB |   1 GB/s | 195 GB/s |

Part of https://github.com/mhmerrill/arkouda/issues/266